Co-locate v1 seeded sampling test
